refactor: PlayedGameResult 패키지 변경
