+++ src/main/java/domain/gameplaying/Participants.java
@@ -1,6 +1,6 @@
 package domain.gameplaying;
 
-import dto.PlayedGameResult;
+import domain.PlayedGameResult;
 import java.util.List;
 
 public class Participants {
