+++ src/main/java/domain/gameplaying/Players.java
@@ -1,6 +1,6 @@
 package domain.gameplaying;
 
-import dto.PlayedGameResult;
+import domain.PlayedGameResult;
 import java.util.ArrayDeque;
 import java.util.Deque;
 import java.util.List;
